Server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paho-mqtt==2.0.0
pyyaml==6.0.1
orjson==3.9.12
pydantic==2.5.3

